KI-Komponente/backend/requirements.txt
fastapi==0.115.11
python-dotenv==1.0.1
requests==2.32.3
flake8==7.1.2
pytest==8.3.5
httpx==0.28.1
sphinx==8.1.3
pillow==11.1.0
uvicorn==0.34.0
pytest-cov==6.0.0
sphinx-rtd-theme==3.0.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
scikit-learn==1.6.1
pandas==2.2.3
matplotlib==3.10.1
numpy==2.2.4